Refactor Contact page into Presentational Components (#123)
