+++ src/features/contact/components/FormField.tsx
@@ -1,0 +1,33 @@
+import React, { useId } from 'react';
+import { Box, Stack, Text } from '@/layouts/Primitives';
+
+interface FormFieldProps {
+  label: string;
+  error?: string;
+  children: React.ReactElement;
+}
+
+export function FormField({ label, error, children }: FormFieldProps) {
+  const id = useId();
+  const errorId = `${id}-error`;
+
+  return (
+    <Stack gap={2} marginBottom={6}>
+      <Box display="flex" justify="between" align="center">
+        <Text as="label" htmlFor={id} variant="mono" size="xs" weight="font-semibold" color="dim" className="tracking-[0.15em] uppercase">
+          {label}
+        </Text>
+        {error && (
+          <Text id={errorId} variant="mono" weight="font-semibold" color="brand" size="xs" role="alert">
+            {error}
+          </Text>
+        )}
+      </Box>
+      {React.cloneElement(children, {
+        id,
+        'aria-describedby': error ? errorId : undefined,
+        'aria-invalid': !!error
+      })}
+    </Stack>
+  );
+}
