+++ src/pages/Contact.tsx
@@ -1,2 +1,40 @@
-import Contact from '@/features/profile/ContactConsole';
-export default Contact;
+import React from 'react';
+import { useContactForm } from '@/hooks/use-contact-form';
+import { SuccessState } from '@/features/contact/components/SuccessState';
+import { ContactFormView } from '@/features/contact/components/ContactFormView';
+
+/**
+ * Contact Page Container
+ * Follows separation of concerns by keeping orchestration logic here
+ * and presentation logic in the feature components.
+ */
+export default function Contact() {
+  const {
+    formData,
+    handleChange,
+    errors,
+    isSubmitting,
+    submitted,
+    submit,
+    reset
+  } = useContactForm();
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    submit();
+  };
+
+  if (submitted) {
+    return <SuccessState onReset={reset} />;
+  }
+
+  return (
+    <ContactFormView
+      formData={formData}
+      errors={errors}
+      isSubmitting={isSubmitting}
+      onChange={handleChange}
+      onSubmit={handleSubmit}
+    />
+  );
+}
